Animation graph: move code
