+++ cocos/animation/marionette/graph-debug.ts
@@ -24,15 +24,13 @@
 
 import { EDITOR } from 'internal:constants';
 import { editorExtrasTag } from '../../core';
-
-import { AnimationBlend } from './animation-blend';
-import { ClipMotion } from './clip-motion';
+import { Motion } from './motion';
 
 export const RUNTIME_ID_ENABLED = EDITOR;
 
 export type RuntimeID = number;
 
-export function getMotionRuntimeID (motion: ClipMotion | AnimationBlend) {
+export function getMotionRuntimeID (motion: Motion) {
     return (motion[editorExtrasTag] as undefined | { id?: RuntimeID })?.id;
 }
