+++ cocos/animation/marionette/motion/animation-blend-direct.ts
@@ -22,14 +22,14 @@
  THE SOFTWARE.
 */
 
-import { _decorator } from '../../core';
-import { createEval } from './create-eval';
+import { _decorator } from '../../../core';
+import { createEval } from '.././create-eval';
 import { MotionEvalContext } from './motion';
 import { AnimationBlend, AnimationBlendEval, AnimationBlendItem } from './animation-blend';
-import { CLASS_NAME_PREFIX_ANIM } from '../define';
-import { ReadonlyClipOverrideMap } from './graph-eval';
-import { BindableNumber, bindOr, VariableType } from './parametric';
-import { AnimationGraphBindingContext } from './animation-graph-context';
+import { CLASS_NAME_PREFIX_ANIM } from '../../define';
+import type { ReadonlyClipOverrideMap } from '.././clip-overriding';
+import { BindableNumber, bindOr, VariableType } from '.././parametric';
+import { AnimationGraphBindingContext } from '.././animation-graph-context';
 
 const { ccclass, serializable } = _decorator;
