+++ cocos/animation/marionette/motion/index.ts
@@ -1,0 +1,7 @@
+export { Motion } from './motion';
+export type { MotionEval, MotionEvalContext, MotionPort } from './motion';
+export { ClipMotion } from './clip-motion';
+export { AnimationBlend } from './animation-blend';
+export { AnimationBlend1D } from './animation-blend-1d';
+export { AnimationBlend2D } from './animation-blend-2d';
+export { AnimationBlendDirect } from './animation-blend-direct';
