+++ cocos/animation/marionette/runtime-exports.ts
@@ -23,10 +23,7 @@
 */
 
 import './animation-graph';
-import './clip-motion';
-import './animation-blend-1d';
-import './animation-blend-2d';
-import './animation-blend-direct';
+import './motion';
 import './animation-mask';
 import './animation-graph-variant';
 
@@ -37,5 +34,5 @@ export type { AnimationGraphVariantRunTime } from './animation-graph-variant';
 export { AnimationController } from './animation-controller';
 export type { ClipStatus, TransitionStatus, MotionStateStatus } from './animation-controller';
 export { VariableType } from './parametric';
-export { StateMachineComponent } from './state-machine-component';
+export { StateMachineComponent } from './state-machine/state-machine-component';
 export type { Value } from './variable';
