+++ cocos/animation/marionette/state-machine/condition/condition-base.ts
@@ -1,0 +1,19 @@
+import { BindContext } from '../../parametric';
+import { _decorator } from '../../../../core';
+import { createEval } from '../../create-eval';
+import { VariableTypeMismatchedError } from '../../errors';
+import { AnimationGraphBindingContext } from '../../animation-graph-context';
+
+export type ConditionEvalContext = BindContext;
+
+export interface Condition {
+    clone (): Condition;
+    [createEval] (context: AnimationGraphBindingContext): ConditionEval;
+}
+
+export interface ConditionEval {
+    /**
+     * Evaluates this condition.
+     */
+    eval(): boolean;
+}
