+++ cocos/animation/marionette/state-machine/state-machine-component.ts
@@ -23,8 +23,8 @@
 */
 
 import { ccclass } from 'cc.decorator';
-import { CLASS_NAME_PREFIX_ANIM } from '../define';
-import type { AnimationController, MotionStateStatus } from './animation-controller';
+import { CLASS_NAME_PREFIX_ANIM } from '../../define';
+import type { AnimationController, MotionStateStatus } from '../animation-controller';
 
 /**
  * @en State machine component.
