+++ tests/animation/animation-graph-asset.test.ts
@@ -1,12 +1,9 @@
 import { Vec2 } from "../../cocos/core";
 import { AnimationClip } from "../../cocos/animation/animation-clip";
-import { AnimationBlend1D } from "../../cocos/animation/marionette/animation-blend-1d";
-import { AnimationBlend2D } from "../../cocos/animation/marionette/animation-blend-2d";
+import { Motion, ClipMotion, AnimationBlend1D, AnimationBlend2D } from "../../cocos/animation/marionette/motion";
 import { AnimationGraph, AnimationTransition, EmptyState, EmptyStateTransition, StateMachine, SubStateMachine, Transition } from "../../cocos/animation/marionette/animation-graph";
-import { ClipMotion } from "../../cocos/animation/marionette/clip-motion";
-import { BinaryCondition, Condition, TriggerCondition, UnaryCondition } from "../../cocos/animation/marionette/condition";
-import { Motion } from "../../cocos/animation/marionette/motion";
-import { MotionState } from "../../cocos/animation/marionette/motion-state";
+import { BinaryCondition, Condition, TriggerCondition, UnaryCondition } from "../../cocos/animation/marionette/state-machine/condition";
+import { MotionState } from "../../cocos/animation/marionette/state-machine/motion-state";
 import { Bindable } from "../../cocos/animation/marionette/parametric";
 import { assertIsTrue } from "../../cocos/core/data/utils/asserts";
