+++ tests/animation/new-gen-anim/utils/node-transform-value-observer.ts
@@ -1,6 +1,6 @@
 import { QuatTrack, VectorTrack } from "../../../../cocos/animation/animation";
 import { AnimationClip } from "../../../../cocos/animation/animation-clip";
-import { ClipMotion } from "../../../../cocos/animation/marionette/clip-motion";
+import { ClipMotion } from "../../../../cocos/animation/marionette/motion";
 import { Quat, toDegree, toRadian, Vec3 } from "../../../../cocos/core";
 import { Node } from "../../../../cocos/scene-graph";
 import { CreateMotionContext } from "./fixtures";
