+++ tests/animation/newgenanim.test.ts
@@ -3,18 +3,19 @@ import { lerp, Vec3, warnID } from '../../cocos/core';
 import { AnimationBlend1D, AnimationBlend2D, Condition, InvalidTransitionError, VariableNotDefinedError, ClipMotion, AnimationBlendDirect, VariableType, AnimationMask, AnimationGraphVariant } from '../../cocos/animation/marionette/asset-creation';
 import { AnimationGraph, StateMachine, Transition, isAnimationTransition, AnimationTransition, TransitionInterruptionSource, State, Layer } from '../../cocos/animation/marionette/animation-graph';
 import { VariableTypeMismatchedError } from '../../cocos/animation/marionette/errors';
-import { AnimationGraphEval, MotionStateStatus, ClipStatus } from '../../cocos/animation/marionette/graph-eval';
-import { blend1D } from '../../cocos/animation/marionette/blend-1d';
+import { AnimationGraphEval } from '../../cocos/animation/marionette/graph-eval';
+import { MotionStateStatus, ClipStatus } from '../../cocos/animation/marionette/state-machine/state-machine-eval';
+import { blend1D } from '../../cocos/animation/marionette/motion/blend-1d';
 import '../utils/matcher-deep-close-to';
-import { BinaryCondition, UnaryCondition, TriggerCondition } from '../../cocos/animation/marionette/condition';
+import { BinaryCondition, UnaryCondition, TriggerCondition } from '../../cocos/animation/marionette/state-machine/condition';
 import { AnimationController } from '../../cocos/animation/marionette/animation-controller';
-import { StateMachineComponent } from '../../cocos/animation/marionette/state-machine-component';
+import { StateMachineComponent } from '../../cocos/animation/marionette/state-machine/state-machine-component';
 import { VectorTrack } from '../../cocos/animation/animation';
 import 'jest-extended';
 import { assertIsTrue } from '../../cocos/core/data/utils/asserts';
 import { AnimationClip } from '../../cocos/animation/animation-clip';
 import { TriggerResetMode } from '../../cocos/animation/marionette/variable';
-import { MotionState } from '../../cocos/animation/marionette/motion-state';
+import { MotionState } from '../../cocos/animation/marionette/state-machine/motion-state';
 import { Node, Component } from '../../cocos/scene-graph';
 import * as maskTestHelper from './new-gen-anim/utils/mask-test-helper';
 import '../utils/matchers/value-type-asymmetric-matchers';
